requirements.txt
gradio>=3.50.2
transformers>=4.30.0
pillow>=9.0.0
numpy>=1.21.0
torch>=2.0.0
easyocr>=1.7.0
opencv-python>=4.5.0
pdf2image>=1.16.3
PyPDF2>=3.0.0
pandas>=1.3.0
openpyxl>=3.0.0
